+++ test/framework/src/main/java/org/elasticsearch/test/ESIntegTestCase.java
@@ -27,7 +27,6 @@ import org.apache.http.HttpHost;
 import org.apache.lucene.search.Sort;
 import org.apache.lucene.util.IOUtils;
 import org.apache.lucene.util.LuceneTestCase;
-import org.apache.lucene.util.TestUtil;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.ExceptionsHelper;
 import org.elasticsearch.action.ActionListener;
@@ -141,7 +140,6 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.lang.annotation.Annotation;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
@@ -151,7 +149,6 @@ import java.lang.annotation.Target;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.URL;
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -2113,48 +2110,7 @@ public abstract class ESIntegTestCase extends ESTestCase {
         return internalCluster().routingKeyForShard(resolveIndex(index), shard, random());
     }
 
-    /**
-     * Return settings that could be used to start a node that has the given zipped home directory.
-     */
-    protected Settings prepareBackwardsDataDir(Path backwardsIndex, Object... settings) throws IOException {
-        Path indexDir = createTempDir();
-        Path dataDir = indexDir.resolve("data");
-        try (InputStream stream = Files.newInputStream(backwardsIndex)) {
-            TestUtil.unzip(stream, indexDir);
-        }
-        assertTrue(Files.exists(dataDir));
 
-        // list clusters in the datapath, ignoring anything from extrasfs
-        final Path[] list;
-        try (DirectoryStream<Path> stream = Files.newDirectoryStream(dataDir)) {
-            List<Path> dirs = new ArrayList<>();
-            for (Path p : stream) {
-                if (!p.getFileName().toString().startsWith("extra")) {
-                    dirs.add(p);
-                }
-            }
-            list = dirs.toArray(new Path[0]);
-        }
-
-        if (list.length != 1) {
-            StringBuilder builder = new StringBuilder("Backwards index must contain exactly one cluster\n");
-            for (Path line : list) {
-                builder.append(line.toString()).append('\n');
-            }
-            throw new IllegalStateException(builder.toString());
-        }
-        Path src = list[0].resolve(NodeEnvironment.NODES_FOLDER);
-        Path dest = dataDir.resolve(NodeEnvironment.NODES_FOLDER);
-        assertTrue(Files.exists(src));
-        Files.move(src, dest);
-        assertFalse(Files.exists(src));
-        assertTrue(Files.exists(dest));
-        Settings.Builder builder = Settings.builder()
-            .put(settings)
-            .put(Environment.PATH_DATA_SETTING.getKey(), dataDir.toAbsolutePath());
-
-        return builder.build();
-    }
 
     @Override
     protected NamedXContentRegistry xContentRegistry() {
